refactor(composer_rpc): extract delivery simulation to shared module
Move simulate_l1_delivery (520 LOC) and its 4 helper functions
(compute_proxy_address_on_l1, extract_delivery_output_from_trigger_trace,
find_delivery_call, extract_l1_to_l2_return_calls) from l2_to_l1.rs
to a new shared delivery.rs module (804 LOC).

These functions are called by both the Direction hook
(enrich_calls_before_retrace) and the post-discovery processing.

l2_to_l1.rs: 4958 → 4170 (-788 lines).
Direction.rs updated to use super::delivery::simulate_l1_delivery.

616/616 tests, clippy clean.
