+++ crates/based-rollup/src/composer_rpc/l2_to_l1.rs
@@ -42,7 +42,7 @@ use super::common::{
     get_l1_block_context as get_l1_block_context_for_proxy,
     get_verification_key as get_verification_key_for_proxy,
 };
-use super::model::{DiscoveredCall, L1ProxyLookup, L2ProxyLookup, ReturnEdge};
+use super::model::{DiscoveredCall, L2ProxyLookup, ReturnEdge};
 
 /// Run the RPC proxy server.
 ///
@@ -783,7 +783,7 @@ async fn queue_independent_calls_l2_to_l1(
 /// This enrichment ensures that the next iteration of `simulate_l1_delivery` builds
 /// inner RESULT entries with real data (instead of `data: vec![]`), so the L1
 /// delivery function (e.g., Logger) receives the correct inner return value (#246).
-async fn enrich_return_calls_via_l2_trace(
+pub(crate) async fn enrich_return_calls_via_l2_trace(
     client: &reqwest::Client,
     l2_rpc_url: &str,
     cross_chain_manager_address: Address,
@@ -1617,543 +1617,9 @@ async fn try_chained_l2_enrichment(
     all_enriched
 }
 
-/// Simulate L1 delivery via `debug_traceCallMany` to capture return data.
-///
-/// For EOA targets (no code on L1), returns `(vec![], false, vec![])` immediately.
-/// For contract targets, builds preliminary L1 deferred entries, signs an ECDSA
-/// proof, and simulates `[postBatch, createProxy, trigger]` on L1 via
-/// `debug_traceCallMany`. Walks the trigger trace to extract the delivery CALL's
-/// return data and success/failure status.
-///
-/// **Iterative discovery**: After each simulation, walks the trigger trace for
-/// L1→L2 return calls (multi-call continuation pattern). If new return calls are found, rebuilds
-/// entries incorporating the return calls as continuations and re-simulates until
-/// convergence or MAX_SIMULATION_ITERATIONS is reached. This mirrors the L1 proxy's
-/// iterative `traceCallMany` loop.
-///
-/// Returns `(return_data, failed, detected_return_calls)`.
-/// Returns `None` if the simulation cannot be performed.
-#[allow(clippy::too_many_arguments)]
-pub(crate) async fn simulate_l1_delivery(
-    client: &reqwest::Client,
-    l1_rpc_url: &str,
-    l2_rpc_url: &str,
-    cross_chain_manager_address: Address,
-    rollups_address: Address,
-    builder_address: Address,
-    builder_private_key: Option<&str>,
-    rollup_id: u64,
-    _trigger_user: Address,
-    destination: Address,
-    data: &[u8],
-    value: U256,
-    rlp_encoded_tx: &[u8],
-    root_scope: &[U256],
-    known_delivery_return_data: &[u8],
-    known_delivery_failed: bool,
-) -> Option<(Vec<u8>, bool, Vec<ReturnEdge>)> {
-    // First check if destination has code on L1.
-    // If it's an EOA, return data is empty and we skip simulation.
-    let code_req = serde_json::json!({
-        "jsonrpc": "2.0",
-        "method": "eth_getCode",
-        "params": [format!("{destination}"), "latest"],
-        "id": 1
-    });
-    let code_resp = client.post(l1_rpc_url).json(&code_req).send().await.ok()?;
-    let code_body: Value = code_resp.json().await.ok()?;
-    let code_hex = code_body.get("result")?.as_str()?;
-    if code_hex == "0x" || code_hex == "0x0" {
-        // EOA target — no return data
-        return Some((vec![], false, vec![]));
-    }
-
-    // Contract target — full traceCallMany simulation.
-    // Build preliminary L1 deferred entries, sign postBatch proof, simulate
-    // [postBatch, executeL2TX] on L1, and extract delivery return data.
-    tracing::info!(
-        target: "based_rollup::proxy",
-        %destination,
-        data_len = data.len(),
-        "L1 target has code — running full traceCallMany simulation"
-    );
-
-    // Parse builder private key — required for signing postBatch proof.
-    let key_hex = builder_private_key?;
-    let key_clean = key_hex.strip_prefix("0x").unwrap_or(key_hex);
-    let builder_key = match key_clean.parse::<alloy_signer_local::PrivateKeySigner>() {
-        Ok(k) => k,
-        Err(e) => {
-            tracing::warn!(
-                target: "based_rollup::proxy",
-                %e,
-                "failed to parse builder private key for L1 simulation"
-            );
-            return Some((vec![], false, vec![]));
-        }
-    };
-
-    // When delivery failure is already known (from iterative L2 discovery's direct L1
-    // simulation), skip the full postBatch+executeL2TX bundle simulation entirely.
-    // The bundle simulation produces wrong return data for failed deliveries because
-    // placeholder state deltas cause _findAndApplyExecution to miss entries (stateRoot
-    // != currentState), making executeL2TX revert with ExecutionNotFound.
-    // The direct L1 simulation (debug_traceCallMany to destination) already captured
-    // the correct revert data — use it as-is.
-    if known_delivery_failed && !known_delivery_return_data.is_empty() {
-        tracing::info!(
-            target: "based_rollup::proxy",
-            %destination,
-            return_data_len = known_delivery_return_data.len(),
-            return_data_hex = %format!("0x{}", hex::encode(&known_delivery_return_data[..known_delivery_return_data.len().min(40)])),
-            "simulate_l1_delivery: skipping bundle simulation — using known delivery failure data"
-        );
-        return Some((known_delivery_return_data.to_vec(), true, vec![]));
-    }
-
-    // Iterative discovery loop: simulate, extract return calls, rebuild entries, repeat.
-    // Seed with known delivery data from the L2 iterative discovery's own L1 simulation.
-    // This gives correct RESULT hashes on iteration 1, avoiding the convergence cycle
-    // (empty → real → confirm) for calls where delivery_return_data is already known.
-    let mut all_return_calls: Vec<ReturnEdge> = Vec::new();
-    let mut final_return_data: Vec<u8> = known_delivery_return_data.to_vec();
-    let mut prev_return_data: Vec<u8> = known_delivery_return_data.to_vec();
-    let mut final_delivery_failed = known_delivery_failed;
-
-    tracing::info!(
-        target: "based_rollup::proxy",
-        %destination,
-        known_delivery_return_data_len = known_delivery_return_data.len(),
-        known_delivery_return_data_hex = %format!("0x{}", hex::encode(&known_delivery_return_data[..known_delivery_return_data.len().min(40)])),
-        known_delivery_failed,
-        "simulate_l1_delivery: starting with known delivery data"
-    );
-
-    for iteration in 1..=MAX_SIMULATION_ITERATIONS {
-        tracing::info!(
-            target: "based_rollup::proxy",
-            iteration,
-            known_return_calls = all_return_calls.len(),
-            "L1 delivery simulation iteration"
-        );
-
-        // Build L1 deferred entries. On first iteration, use simple L2TX+RESULT entries.
-        // On subsequent iterations, include continuation entries for discovered return calls.
-        let entries = if all_return_calls.is_empty() {
-            // Simple case: just the original L2→L1 call
-            let call_entries = crate::cross_chain::build_l2_to_l1_call_entries(
-                destination,
-                data.to_vec(),
-                value,
-                _trigger_user,
-                rollup_id,
-                rlp_encoded_tx.to_vec(), // RLP-encoded L2 tx for L2TX trigger
-                final_return_data.clone(), // use known data (from iterative discovery or previous iteration)
-                final_delivery_failed,     // use known failed flag
-                root_scope.to_vec(),       // l1_delivery_scope from trace depth
-                crate::cross_chain::TxOutcome::Success,                     // tx_reverts (simulation path, not real queueing)
-            );
-            call_entries.l1_deferred_entries
-        } else {
-            // Continuation case: use the SAME table builder functions as the real
-            // batch (build_l2_to_l1_continuation_entries) to ensure identical entry
-            // ordering. Manual construction produced [L2TX, scope_RESULT, child...]
-            // while the real batch produces [trigger, child..., scope_RESULT], causing
-            // swap-and-pop to consume entries in wrong order during simulation.
-            let root_call = crate::table_builder::L2DetectedCall {
-                destination,
-                data: data.to_vec(),
-                value,
-                source_address: _trigger_user,
-                delivery_return_data: final_return_data.clone(),
-                delivery_failed: final_delivery_failed,
-                scope: ScopePath::from_parts(root_scope.to_vec()),
-                in_reverted_frame: false,
-            };
-
-            let return_calls_for_builder: Vec<crate::table_builder::L2ReturnCall> =
-                all_return_calls
-                    .iter()
-                    .map(|rc| crate::table_builder::L2ReturnCall {
-                        destination: rc.destination,
-                        data: rc.data.clone(),
-                        value: rc.value,
-                        source_address: rc.source_address,
-                        parent_call_index: rc.parent_call_index,
-                        l2_return_data: rc.return_data.clone(),
-                        l2_delivery_failed: rc.delivery_failed,
-                        scope: rc.scope.clone(),
-                    })
-                    .collect();
-
-            let analyzed = crate::table_builder::analyze_l2_to_l1_continuation_calls(
-                &[root_call],
-                &return_calls_for_builder,
-                rollup_id,
-            );
-            let continuation = crate::table_builder::build_l2_to_l1_continuation_entries(
-                &analyzed,
-                crate::cross_chain::RollupId::new(alloy_primitives::U256::from(rollup_id)),
-                rlp_encoded_tx,
-                crate::cross_chain::TxOutcome::Success, // tx_reverts
-            );
-
-            tracing::info!(
-                target: "based_rollup::proxy",
-                l1_entry_count = continuation.l1_entries.len(),
-                return_call_count = all_return_calls.len(),
-                "built simulation entries via table builder (same path as real batch)"
-            );
-
-            continuation.l1_entries
-        };
-
-        // Clear placeholder state deltas for simulation-only entries.
-        // Entries are built with currentState=0x0 / newState=0x0 (placeholders
-        // for the driver to fill with real intermediate roots). In the simulation,
-        // Rollups.sol._findAndApplyExecution checks delta.currentState against the
-        // on-chain root — 0x0 never matches. Clearing deltas makes the match
-        // unconditional (empty deltas → no state root check).
-        // Same fix as L1→L2 path (commit 6885cf0).
-        let mut entries = entries;
-        for e in &mut entries {
-            e.state_deltas.clear();
-        }
-
-        if entries.is_empty() {
-            tracing::warn!(
-                target: "based_rollup::proxy",
-                "entry building produced no L1 deferred entries"
-            );
-            return Some((final_return_data, final_delivery_failed, all_return_calls));
-        }
-
-        // Get L1 block context for proof signing.
-        let (block_number, block_hash, _parent_hash) =
-            match get_l1_block_context_for_proxy(client, l1_rpc_url).await {
-                Ok(ctx) => ctx,
-                Err(e) => {
-                    tracing::warn!(
-                        target: "based_rollup::proxy",
-                        %e,
-                        "failed to get L1 block context for simulation"
-                    );
-                    return Some((final_return_data, final_delivery_failed, all_return_calls));
-                }
-            };
-
-        let trace_block_number = block_number + 1;
-        let trace_parent_hash = block_hash;
-        // For traceCallMany simulation, we control the block timestamp via blockOverride.
-        // Use current time — the override ensures consistency between signed proof and simulation.
-        // Fallback to 0 on the vanishingly rare SystemTime-before-epoch case.
-        let trace_block_timestamp = std::time::SystemTime::now()
-            .duration_since(std::time::UNIX_EPOCH)
-            .map(|d| d.as_secs())
-            .unwrap_or(0);
-
-        // Get verification key from Rollups contract.
-        let vk =
-            match get_verification_key_for_proxy(client, l1_rpc_url, rollups_address, rollup_id)
-                .await
-            {
-                Ok(v) => v,
-                Err(e) => {
-                    tracing::warn!(
-                        target: "based_rollup::proxy",
-                        %e,
-                        "failed to get verification key for simulation"
-                    );
-                    return Some((final_return_data, final_delivery_failed, all_return_calls));
-                }
-            };
-
-        // Sign ECDSA proof for postBatch.
-        let call_data_bytes = alloy_primitives::Bytes::new();
-        let entry_hashes = crate::cross_chain::compute_entry_hashes(&entries, vk);
-        let public_inputs_hash = crate::cross_chain::compute_public_inputs_hash(
-            &entry_hashes,
-            &call_data_bytes,
-            trace_parent_hash,
-            trace_block_timestamp,
-        );
-
-        use alloy_signer::SignerSync;
-        let sig = match builder_key.sign_hash_sync(&public_inputs_hash) {
-            Ok(s) => s,
-            Err(e) => {
-                tracing::warn!(
-                    target: "based_rollup::proxy",
-                    %e,
-                    "failed to sign proof for L1 simulation"
-                );
-                return Some((final_return_data, final_delivery_failed, all_return_calls));
-            }
-        };
-        let sig_bytes = sig.as_bytes();
-        let mut proof_bytes = sig_bytes.to_vec();
-        if proof_bytes.len() == 65 && proof_bytes[64] < 27 {
-            proof_bytes[64] += 27;
-        }
-        let proof = alloy_primitives::Bytes::from(proof_bytes);
-
-        // Encode postBatch calldata.
-        let post_batch_calldata =
-            crate::cross_chain::encode_post_batch_calldata(&entries, call_data_bytes, proof);
-
-        // Encode executeL2TX calldata using typed ABI encoding (NEVER hardcode selectors).
-        let execute_l2tx_calldata = crate::cross_chain::IRollups::executeL2TXCall {
-            rollupId: alloy_primitives::U256::from(rollup_id),
-            rlpEncodedTx: rlp_encoded_tx.to_vec().into(),
-        }
-        .abi_encode();
-
-        // Build traceCallMany request: [postBatch, executeL2TX] in a single bundle.
-        let builder_addr_hex = format!("{builder_address}");
-        let rollups_hex = format!("{rollups_address}");
-        let post_batch_data = format!("0x{}", hex::encode(post_batch_calldata.as_ref()));
-        let execute_l2tx_data = format!("0x{}", hex::encode(&execute_l2tx_calldata));
-        let next_block = format!("{:#x}", trace_block_number);
-
-        let trace_req = serde_json::json!({
-            "jsonrpc": "2.0",
-            "method": "debug_traceCallMany",
-            "params": [
-                [
-                    {
-                        "transactions": [
-                            {
-                                "from": builder_addr_hex,
-                                "to": rollups_hex,
-                                "data": post_batch_data,
-                                "gas": "0x1c9c380"
-                            },
-                            {
-                                "from": builder_addr_hex,
-                                "to": rollups_hex,
-                                "data": execute_l2tx_data,
-                                "gas": "0xc35000"
-                            }
-                        ],
-                        "blockOverride": {
-                            "number": next_block,
-                            "time": format!("{:#x}", trace_block_timestamp)
-                        }
-                    }
-                ],
-                null,
-                { "tracer": "callTracer" }
-            ],
-            "id": 4
-        });
-
-        let resp = match client.post(l1_rpc_url).json(&trace_req).send().await {
-            Ok(r) => match r.json::<Value>().await {
-                Ok(v) => v,
-                Err(e) => {
-                    tracing::warn!(
-                        target: "based_rollup::proxy",
-                        %e,
-                        "traceCallMany response parse failed"
-                    );
-                    return Some((final_return_data, final_delivery_failed, all_return_calls));
-                }
-            },
-            Err(e) => {
-                tracing::warn!(
-                    target: "based_rollup::proxy",
-                    %e,
-                    "traceCallMany request failed"
-                );
-                return Some((final_return_data, final_delivery_failed, all_return_calls));
-            }
-        };
-
-        // Extract traces from result — now 2 traces: [postBatch, executeL2TX].
-        let bundle_traces = match resp
-            .get("result")
-            .and_then(|r| r.get(0))
-            .and_then(|b| b.as_array())
-        {
-            Some(arr) if arr.len() >= 2 => arr,
-            _ => {
-                if let Some(error) = resp.get("error") {
-                    tracing::warn!(
-                        target: "based_rollup::proxy",
-                        ?error,
-                        "traceCallMany returned error"
-                    );
-                } else {
-                    tracing::warn!(
-                        target: "based_rollup::proxy",
-                        "traceCallMany returned unexpected structure (expected 2 traces)"
-                    );
-                }
-                return Some((final_return_data, final_delivery_failed, all_return_calls));
-            }
-        };
-
-        // Check postBatch result (tx0). Log if it reverted but DO NOT bail out.
-        // The trigger trace (tx1) still contains inner calls even when postBatch
-        // reverts (e.g., proof signature mismatch). These inner calls are essential
-        // for depth-N return call detection — the standalone fallback that was here
-        // previously called destination.call() directly (without entries/scope
-        // navigation), which broke all depth-2+ patterns by missing return calls.
-        let tx0_trace = &bundle_traces[0];
-        if tx0_trace.get("error").is_some() || tx0_trace.get("revertReason").is_some() {
-            let error_msg = tx0_trace
-                .get("error")
-                .and_then(|v| v.as_str())
-                .unwrap_or("unknown");
-            tracing::warn!(
-                target: "based_rollup::proxy",
-                error = error_msg,
-                iteration,
-                "postBatch reverted in L1 delivery simulation — \
-                 continuing with trigger trace for return call detection"
-            );
-            // Fall through to trigger trace extraction below.
-            // The trigger tx (tx1) also reverts, but callTracer still captures
-            // all inner calls (proxy calls, executeCrossChainCall, etc.) which
-            // are needed for depth-N return call discovery.
-        }
-
-        // Extract delivery output from executeL2TX trace (tx1).
-        let trigger_trace = &bundle_traces[1];
-        let (return_data, _delivery_failed) =
-            extract_delivery_output_from_trigger_trace(trigger_trace, destination);
-
-        // Trigger simulation is unreliable for L2→L1 calls: entries have
-        // placeholder state deltas and the ECDSA proof may not match real L1
-        // state. When delivery failure was already detected by the caller
-        // (known_delivery_failed=true via iterative L2 discovery), preserve it.
-        // Otherwise assume delivery succeeds — §4f + rewind handles real
-        // failures on L1.
-        if !known_delivery_failed {
-            final_delivery_failed = false;
-        }
-
-        // Extract L1→L2 return calls from the trigger trace BEFORE deciding
-        // the return data fallback — we need to know whether this is depth-1
-        // (no return calls) or depth-2+ (has return calls) to choose the
-        // correct strategy.
-        let new_return_calls = extract_l1_to_l2_return_calls(
-            client,
-            l1_rpc_url,
-            rollups_address,
-            trigger_trace,
-            rollup_id,
-            root_scope,
-        )
-        .await;
-
-        // Filter out already-known return calls using count-based comparison.
-        // Supports legitimate duplicate return calls with identical
-        // (destination, data, value, source_address) tuples. The CALL action hash
-        // includes value and sourceAddress, so two calls with different ETH values
-        // are distinct even if destination and data match.
-        let truly_new = filter_new_by_count(new_return_calls, &all_return_calls, |a, b| {
-            a.destination == b.destination
-                && a.data == b.data
-                && a.value == b.value
-                && a.source_address == b.source_address
-        });
-
-        // Delivery return data from the trigger trace.
-        // Per §C.6, the L2TX terminal RESULT is always void — we don't need
-        // delivery_return_data for the terminal. But we still capture it for
-        // the action_hash of the scope resolution entry (which matches what
-        // _processCallAtScope produces on L1).
-        // The trigger trace may be reverted (placeholder entries), so
-        // find_delivery_call extracts data from inner calls even in reverted frames.
-        final_return_data = return_data.clone();
-
-        // Check convergence: no new return calls AND return data stabilized.
-        // The return data may change across iterations: iteration N captures
-        // revert data (delivery fails without entries for return calls),
-        // iteration N+1 captures success data (entries now present). We need
-        // one more iteration after return data changes so the entries are
-        // rebuilt with the correct RESULT hash (#238).
-        let return_data_changed = final_return_data != prev_return_data;
-        if truly_new.is_empty() && !return_data_changed {
-            tracing::info!(
-                target: "based_rollup::proxy",
-                iteration,
-                total_return_calls = all_return_calls.len(),
-                return_data_len = final_return_data.len(),
-                delivery_hex = %format!("0x{}", hex::encode(&final_return_data)),
-                "L1 delivery simulation converged — no new return calls, return data stable"
-            );
-            // Log each return call's return_data for hash comparison
-            for (ri, rc) in all_return_calls.iter().enumerate() {
-                tracing::info!(
-                    target: "based_rollup::proxy",
-                    ri,
-                    dest = %rc.destination,
-                    return_data_hex = %format!("0x{}", hex::encode(&rc.return_data)),
-                    return_data_len = rc.return_data.len(),
-                    delivery_failed = rc.delivery_failed,
-                    "return call return_data at convergence"
-                );
-            }
-            break;
-        }
-        if truly_new.is_empty() && return_data_changed {
-            tracing::info!(
-                target: "based_rollup::proxy",
-                iteration,
-                old_len = prev_return_data.len(),
-                new_len = final_return_data.len(),
-                "L1 delivery simulation: no new return calls but return data changed — re-iterating for correct RESULT hash"
-            );
-            prev_return_data = final_return_data.clone();
-            continue;
-        }
-
-        tracing::info!(
-            target: "based_rollup::proxy",
-            iteration,
-            new_return_calls = truly_new.len(),
-            "discovered new L1->L2 return calls in delivery trace — re-simulating"
-        );
-
-        all_return_calls.extend(truly_new);
-
-        // Enrich return calls with L2 return data via debug_traceCallMany.
-        // This data is used by the next iteration's simulation bundle so the
-        // inner RESULT entries carry real data instead of void (#246).
-        enrich_return_calls_via_l2_trace(
-            client,
-            l2_rpc_url,
-            cross_chain_manager_address,
-            &mut all_return_calls,
-            rollup_id,
-        )
-        .await;
-
-        prev_return_data = final_return_data.clone();
-    }
-
-    if !all_return_calls.is_empty() {
-        tracing::info!(
-            target: "based_rollup::proxy",
-            count = all_return_calls.len(),
-            return_data_len = final_return_data.len(),
-            delivery_failed = final_delivery_failed,
-            "L1 delivery simulation complete (multi-call continuation pattern detected)"
-        );
-    } else {
-        tracing::info!(
-            target: "based_rollup::proxy",
-            %destination,
-            return_data_len = final_return_data.len(),
-            delivery_failed = final_delivery_failed,
-            "L1 delivery simulation complete"
-        );
-    }
-
-    Some((final_return_data, final_delivery_failed, all_return_calls))
-}
+// `simulate_l1_delivery` moved to `super::delivery::simulate_l1_delivery`.
+// Re-export for backward compatibility with internal callers.
+pub(crate) use super::delivery::simulate_l1_delivery;
 
 /// Simulate N identical cross-chain calls with state accumulation on L1.
 ///
@@ -3022,269 +2488,15 @@ async fn simulate_l1_combined_delivery(
     Some(results)
 }
 
-/// Compute the CrossChainProxy address on L1 for a given trigger user.
-///
-/// Calls `computeCrossChainProxyAddress(originalAddress, originalRollupId)`
-/// on the Rollups contract.
-async fn compute_proxy_address_on_l1(
-    client: &reqwest::Client,
-    l1_rpc_url: &str,
-    rollups_address: Address,
-    trigger_user: Address,
-    rollup_id: u64,
-) -> eyre::Result<Address> {
-    // Encode computeCrossChainProxyAddress(address, uint256)
-    use alloy_sol_types::SolCall;
-    let compute_data = crate::cross_chain::IRollups::computeCrossChainProxyAddressCall {
-        originalAddress: trigger_user,
-        originalRollupId: U256::from(rollup_id),
-    }
-    .abi_encode();
-    let compute_hex = format!("0x{}", hex::encode(&compute_data));
+// `compute_proxy_address_on_l1` moved to `super::delivery`.
+use super::delivery::compute_proxy_address_on_l1;
 
-    let req = serde_json::json!({
-        "jsonrpc": "2.0",
-        "method": "eth_call",
-        "params": [{"to": format!("{rollups_address}"), "data": compute_hex}, "latest"],
-        "id": 99994
-    });
-
-    let resp = client
-        .post(l1_rpc_url)
-        .json(&req)
-        .send()
-        .await?
-        .json::<Value>()
-        .await?;
-
-    if let Some(error) = resp.get("error") {
-        return Err(eyre::eyre!("computeCrossChainProxyAddress failed: {error}"));
-    }
-
-    let result_hex = resp
-        .get("result")
-        .and_then(|v| v.as_str())
-        .ok_or_else(|| eyre::eyre!("no result from computeCrossChainProxyAddress"))?;
-
-    let hex_clean = result_hex.strip_prefix("0x").unwrap_or(result_hex);
-    if hex_clean.len() < 64 {
-        return Err(eyre::eyre!("proxy address return too short"));
-    }
-    let addr_bytes = hex::decode(&hex_clean[..64])
-        .map_err(|_| eyre::eyre!("invalid hex in proxy address return"))?;
-    if addr_bytes.len() < 32 {
-        return Err(eyre::eyre!("proxy address bytes too short"));
-    }
-    Ok(Address::from_slice(&addr_bytes[12..32]))
-}
-
-/// Walk the trigger trace to extract the delivery call's return data and success status.
-///
-/// The trigger trace call tree is:
-/// ```text
-/// proxy fallback → Rollups.executeCrossChainCall → _processCallAtScope → CALL(destination, data, value)
-/// ```
-/// We find the deepest CALL to `destination` and extract its output.
-fn extract_delivery_output_from_trigger_trace(
-    trigger_trace: &Value,
-    destination: Address,
-) -> (Vec<u8>, bool) {
-    let top_level_reverted =
-        trigger_trace.get("error").is_some() || trigger_trace.get("revertReason").is_some();
-
-    if top_level_reverted {
-        let error = trigger_trace
-            .get("error")
-            .and_then(|v| v.as_str())
-            .unwrap_or("unknown");
-        tracing::warn!(
-            target: "based_rollup::proxy",
-            error,
-            "trigger tx reverted in L1 simulation"
-        );
-    }
-
-    // Walk the trace tree looking for the delivery call to destination.
-    // Even when the top-level trigger reverts (common for L2→L1 simulations
-    // with placeholder entries/proofs), the INNER delivery call may have
-    // succeeded and its output is visible in the callTracer subcalls.
-    // The output is critical: it becomes the RESULT entry's data field,
-    // and a mismatch (void vs actual return data) causes ExecutionNotFound.
-    if let Some((output, failed)) = find_delivery_call(trigger_trace, destination) {
-        if failed {
-            // The delivery call itself reverted. This happens when the destination
-            // function triggers further cross-chain calls (non-leaf delivery) and
-            // the entries for those calls aren't loaded yet — the inner CCM
-            // `executeCrossChainCall` reverts with ExecutionNotFound, propagating
-            // up through the proxy to the delivery call.
-            //
-            // The `output` here is the error selector (e.g., 0xed6bc750 =
-            // ExecutionNotFound), NOT the function's actual return data. Using it
-            // as RESULT entry data produces a wrong hash, forcing 2 extra
-            // convergence iterations (error→void→confirm).
-            //
-            // Return void instead: all non-leaf deliveries in practice return void
-            // (incrementProxy, deepCall, callBoth, etc.). If a future contract has
-            // a non-leaf delivery returning non-void, the RESULT hash mismatch
-            // triggers ExecutionNotFound on L1 → rewind (existing safety net).
-            //
-            // Evidence: 16/16 non-leaf delivery successes across 14 E2E tests
-            // return void (0 bytes). 0/16 return non-void.
-            return (vec![], false);
-        }
-        return (output, failed);
-    }
-
-    // Only return revert if we truly couldn't find any delivery output.
-    if top_level_reverted {
-        return (vec![], true);
-    }
-
-    // If we couldn't find the delivery call in the trace, return empty.
-    tracing::info!(
-        target: "based_rollup::proxy",
-        %destination,
-        "could not find delivery call to destination in trigger trace — returning empty"
-    );
-    (vec![], false)
-}
-
-/// Recursively search the call trace for a CALL to the given destination address.
-/// Returns the deepest match's (output_bytes, failed).
-fn find_delivery_call(trace: &Value, destination: Address) -> Option<(Vec<u8>, bool)> {
-    let dest_lower = format!("{destination}").to_lowercase();
-
-    // Check subcalls first (depth-first) to find the deepest match.
-    if let Some(calls) = trace.get("calls").and_then(|v| v.as_array()) {
-        for subcall in calls {
-            if let Some(result) = find_delivery_call(subcall, destination) {
-                return Some(result);
-            }
-        }
-    }
-
-    // Check this node.
-    let to = trace.get("to").and_then(|v| v.as_str()).unwrap_or("");
-    if to.to_lowercase() == dest_lower {
-        let failed = trace.get("error").is_some();
-        let output_hex = trace.get("output").and_then(|v| v.as_str()).unwrap_or("0x");
-        let hex_clean = output_hex.strip_prefix("0x").unwrap_or(output_hex);
-        let output_bytes = hex::decode(hex_clean).unwrap_or_default();
-        return Some((output_bytes, failed));
-    }
-
-    None
-}
-
-// `ReturnEdge` was replaced by `ReturnEdge` from `super::model`.
-// See the import at the top of this file.
-
-/// Maximum number of iterative discovery rounds in `simulate_l1_delivery`.
-const MAX_SIMULATION_ITERATIONS: usize = 10;
-
-/// Extract L1→L2 return calls from the trigger trace using the generic
-/// `trace::walk_trace_tree` with ephemeral proxy support.
-///
-/// Walks the trigger trace depth-first via the protocol-level detection
-/// (`executeCrossChainCall` child pattern on Rollups.sol). Filters results
-/// to only include return calls targeting our rollup. Manager-originated
-/// calls (forward delivery by Rollups) are automatically skipped by the
-/// generic walker.
-///
-/// Uses `authorizedProxies` for persistent proxies AND scans the trace for
-/// `createCrossChainProxy` to detect ephemeral proxies (created during the
-/// trigger execution).
-async fn extract_l1_to_l2_return_calls(
-    client: &reqwest::Client,
-    l1_rpc_url: &str,
-    rollups_address: Address,
-    trigger_trace: &Value,
-    our_rollup_id: u64,
-    parent_scope: &[U256],
-) -> Vec<ReturnEdge> {
-    let lookup = L1ProxyLookup {
-        client,
-        rpc_url: l1_rpc_url,
-        rollups_address,
-    };
-    let mut proxy_cache: std::collections::HashMap<Address, Option<super::trace::ProxyInfo>> =
-        std::collections::HashMap::new();
-    let mut ephemeral_proxies = std::collections::HashMap::new();
-    let mut detected_calls = Vec::new();
-
-    // Rollups.sol is the manager contract on L1.
-    super::trace::walk_trace_tree(
-        trigger_trace,
-        &[rollups_address],
-        &lookup,
-        &mut proxy_cache,
-        &mut ephemeral_proxies,
-        &mut detected_calls,
-        &mut std::collections::HashSet::new(),
-    )
-    .await;
-
-    // Convert trace::DetectedCall to ReturnEdge, filtering to only
-    // include return calls targeting our rollup.
-    //
-    // walk_trace_tree returns ALL proxy calls (both the original L2→L1 trigger
-    // and any L1→L2 return calls). We filter by rollup_id to keep only calls
-    // that target our rollup (return calls) and exclude calls targeting L1 or
-    // other rollups (forward calls).
-    //
-    // The walker already skips manager-originated calls (from=Rollups), so
-    // forward delivery calls (Rollups calling proxy.executeOnBehalf) are not
-    // in the results.
-    detected_calls
-        .into_iter()
-        .filter_map(|c| {
-            // Recover rollup_id from the proxy cache or ephemeral proxies.
-            // The walker resolved proxy identity for detection — look it up
-            // by matching original_address == c.destination.
-            let proxy_info = proxy_cache
-                .values()
-                .find_map(|opt| opt.filter(|info| info.original_address == c.destination))
-                .or_else(|| {
-                    ephemeral_proxies
-                        .values()
-                        .find(|info| info.original_address == c.destination)
-                        .copied()
-                });
-
-            match proxy_info {
-                Some(info) if info.original_rollup_id == our_rollup_id => {
-                    tracing::info!(
-                        target: "based_rollup::proxy",
-                        dest = %c.destination,
-                        source = %c.source_address,
-                        data_len = c.calldata.len(),
-                        value = %c.value,
-                        rollup_id = info.original_rollup_id,
-                        "detected L1->L2 return call in delivery trace via walk_trace_tree"
-                    );
-                    // Accumulated scope: parent's scope ++ [0].
-                    // Return calls are always first-children of their triggering scope
-                    // (per reentrantCrossChainCalls E2E spec). The trace_depth from the
-                    // L1 trigger trace is NOT used because it includes protocol-internal
-                    // frames (Rollups.newScope, executeOnBehalf) that inflate the depth.
-                    let mut accumulated = parent_scope.to_vec();
-                    accumulated.push(U256::ZERO);
-                    Some(ReturnEdge {
-                        destination: c.destination,
-                        data: c.calldata,
-                        value: c.value,
-                        source_address: c.source_address,
-                        parent_call_index: crate::cross_chain::ParentLink::Root,
-                        return_data: vec![],
-                        delivery_failed: false,
-                        scope: ScopePath::from_parts(accumulated),
-                    })
-                }
-                _ => None, // Not targeting our rollup — skip (forward call or other rollup)
-            }
-        })
-        .collect()
-}
+// `extract_delivery_output_from_trigger_trace`, `find_delivery_call`,
+// `MAX_SIMULATION_ITERATIONS`, and `extract_l1_to_l2_return_calls`
+// moved to `super::delivery`. Re-export for internal callers.
+use super::delivery::extract_delivery_output_from_trigger_trace;
+use super::delivery::extract_l1_to_l2_return_calls;
+use super::delivery::MAX_SIMULATION_ITERATIONS;
 
 /// Build a JSON-RPC response with a gas estimate.
 fn build_gas_estimate_response(request_json: &Value, gas: u64) -> String {
